Updated project builder to validate the start URL, removed getW3cManifest function from manifest tools and added new validateAndNormalizeStartUrl function.
