+++ lib/manifestTools.js
@@ -63,51 +63,6 @@ function convertTo(manifestInfo, outputFormat, callback) {
   });
 }
 
-function getW3cManifest(siteUrl, manifestLocation, callback) {
-    var parsedSiteUrl = url.parse(siteUrl);
-
-    function manifestRetrieved(err, manifestInfo) {
-        if (err) {
-            return callback(err, manifestInfo);
-        }
-
-        if (manifestInfo.format !== c.BASE_MANIFEST_FORMAT) {
-            return callback(new Error("The manifest found is not a W3C manifest."), manifestInfo);
-        }
-
-        var parsedManifestStartUrl = url.parse(manifestInfo.content.start_url);
-        if (parsedManifestStartUrl.hostname && parsedSiteUrl.hostname !== parsedManifestStartUrl.hostname) {
-            return callback(new Error("The domain of the hosted site (" + parsedSiteUrl.hostname + ") does not match the domain of the manifest's start_url parameter (" + parsedManifestStartUrl.hostname + ")"), manifestInfo);
-        }
-
-        // make sure the manifest's start_url is an absolute URL
-        manifestInfo.content.start_url = url.resolve(siteUrl, manifestInfo.content.start_url);
-
-        return callback(undefined, manifestInfo);
-    }
-
-    if (!parsedSiteUrl.hostname) {
-        return callback(new Error("The site URL is not a valid URL."));
-    }
-
-    if (manifestLocation) {
-        var parsedManifestUrl = url.parse(manifestLocation);
-        if (parsedManifestUrl && parsedManifestUrl.host) {
-            // download manifest from remote location
-            log.info('Downloading manifest from ' + manifestLocation + '...');
-            downloadManifestFromUrl(manifestLocation, manifestRetrieved);
-        } else {
-            // read local manifest file
-            log.info('Reading manifest file ' + manifestLocation + '...');
-            getManifestFromFile(manifestLocation, manifestRetrieved);
-        }
-    } else {
-        // scan a site to retrieve its manifest
-        log.info('Scanning ' + siteUrl + ' for manifest...');
-        getManifestFromSite(siteUrl, manifestRetrieved);
-    }
-}
-
 function getManifestUrlFromSite(siteUrl, callback) {
   request({
     uri: siteUrl
@@ -279,15 +234,31 @@ function validateManifest(manifestInfo, targetPlatforms, callback) {
       });
 }
 
-module.exports = {
-  getW3cManifest: getW3cManifest,
+function validateAndNormalizeStartUrl(siteUrl, manifestInfo, callback) {
+  if (manifestInfo.format !== c.BASE_MANIFEST_FORMAT) {
+    return callback(new Error("The manifest found is not a W3C manifest."), manifestInfo);
+  }
 
+  var parsedManifestStartUrl = url.parse(manifestInfo.content.start_url);
+  if (parsedManifestStartUrl.hostname && parsedSiteUrl.hostname !== parsedManifestStartUrl.hostname) {
+    return callback(new Error("The domain of the hosted site (" + parsedSiteUrl.hostname + ") does not match the domain of the manifest's start_url parameter (" + parsedManifestStartUrl.hostname + ")"), manifestInfo);
+  }
+
+  // make sure the manifest's start_url is an absolute URL
+  manifestInfo.content.start_url = url.resolve(siteUrl, manifestInfo.content.start_url);
+
+  return callback(undefined, manifestInfo);
+}
+
+module.exports = {
   getManifestFromSite: getManifestFromSite,
   getManifestFromFile: getManifestFromFile,
   writeToFile: writeToFile,
 
   getManifestUrlFromSite: getManifestUrlFromSite,
   downloadManifestFromUrl: downloadManifestFromUrl,
+
+  validateAndNormalizeStartUrl: validateAndNormalizeStartUrl,
 
   convertTo: convertTo,
