+++ lib/projectBuilder.js
@@ -9,9 +9,25 @@ var manifestTools = require('./manifestTools'),
     log = require('loglevel'),
     downloader = require('./projectBuilder/downloader'),
     Q = require('q'),
+    c = require('./manifestTools/constants'),
     mkdirp = require('mkdirp');
 
+function validateManifest(manifestInfo) {
+  if (manifestInfo.format !== c.BASE_MANIFEST_FORMAT) {
+    return new Error("The manifest passed as parameter is not a W3C manifest.");
+  }
+
+  var parsedManifestStartUrl = url.parse(manifestInfo.content.start_url);
+  if (!parsedManifestStartUrl.hostname) {
+    return new Error("The manifest's start_url parameter (" + parsedManifestStartUrl.hostname + ") is not a full url.");
+  }
+}
+
 var createApps = function (w3cManifestInfo, rootDir, platforms, build, callback) {
+  var err = validateManifest(w3cManifestInfo);
+  if (err) {
+    return callback(err);
+  }
 
   // determine the path where the Cordova app will be created
   var appName = w3cManifestInfo.content.short_name;
@@ -23,7 +39,7 @@ var createApps = function (w3cManifestInfo, rootDir, platforms, build, callback)
       return callback(err);
     }
 
-    // process all requested platforms    
+    // process all requested platforms
     var pendingTasks = [];
     var cordovaPlatforms = [];
     platforms.forEach(function (el) {
@@ -164,6 +180,11 @@ var createFirefoxApp = function (w3cManifestInfo, generatedAppDir, build) {
 };
 
 var createCordovaApp = function (w3cManifestInfo, generatedAppDir, platforms, build) {
+  var err = validateManifest(w3cManifestInfo);
+  if (err) {
+    return callback(err);
+  }
+
   log.info('Generating the Cordova application...');
 
   var task = Q.defer();
